Moved from stats-html. Changed maxworkrows to always be 0 (multiply by 0) set nocount off for the final query (php depends on this for it's rowcount info)
This needs to eventually be made into a contest generic script (s/csc/${1}/g) and
given a .sh script to create the csc and rc5_64 versions.

This also needs to be broken apart some. The excelent escapewildcards function should
really be it's own proc. I think that the first query can also get poorly optimized
depending on what the first query run against the proc is.

NOTE: This version has a bug in it. Search on chris@open and then search on chris@.
You will notice that chris@open does not show up in the second search.