Initical commit.
I've taken the liberty of taking Bruce's csc_psearch code and 'dBizing' it.
It's now setup just like the lastupdate stuff, where the actual procedure
skeletons live in the psearch directory, and psearch.sh adds procedures for
all the contests. I've also segregated Bruce's original procedure into a
main procedure (p_psearch) and contest specific procedures.

usage:"
	p_psearch contest, searchtext, maxrows, escapewildcards"
where:"
	contest is contest to look at"
	searchtext is the email to search for"
	maxrows is the maximum number of rows to return"
	escapewildcards is set to 1 if you want to escape wildcards in searchtext"
		(%, -) [default], or 0 if you do not want to escape them"

I think I've come to the conclusiong that the split procedure architecture is
better than embedding contest names in procedure names because it allows you to
do error-checking of the project name in the main proc (such as swapping '-' for
'_', although I'm not currecntly doing this) and it greatly minimizes the project-
specific code (ie: the only thing from the original proc that is in the project-
specific procedure is the final select statement).